ruuvitag_sensor.adapters.nix_hci: Add debug logging, rework parsing
This adds additional debug logging to the ruuvitag_sensor.adapters.nix_hci
adapter, and reworks the parser to validate the data structure better,
and filter out some non-interesting messages early.

It also changes the structure of the class slightly to make it easier
to generate a child class. This will be used later to add an adapter
that reads hcidump data from a file instead.