Add ruuvitag_sensor.adapters.nix_hci_file
This adapter is very similar to nix_hci, but instead reads data from a
file in the same format that hcidump produces. To enable use of this
adapter, set the RUUVI_NIX_FROMFILE environment variable. The value does
not matter.

The file name to read will be passed in in place of the bluetooth
adapter, so any program that allows setting the adapter can instead
use this.