ruuvitag_sensor.data_formats: Rework parsing
This reworks the data parsing in ruuvitag_sensor.data_formats to more
closely follow the structure defined for BLE advertisements. It checks
for types that we're interested in for the different firmware versions,
and returns the appropriate data.

This will give us more confidence that the data we found is actually
good.

It also allows for dealing with non-measurement advertisements sent from
tags with firmware 3.x.