Reliability improvement to the ReactiveX + InfluxDB example
If the InfluxDB write fails for some reason (e.g. intermittent DNS resolution issue)
and the exception is not caught, the process reading data from a specific Ruuvitag gets
killed and no more data will be gathered from said tag. As the exception does not kill
the whole program the error might go unnoticed. Catching the error and printing it out
means that the program will keep on trying to write new data as it comes, but the issue
will also be logged.