Fix a Bleak deprecation warning
The following warning is displayed when Bleak v0.18.0 is used:
ruuvitag-sensor/ruuvitag_sensor/adapters/bleak_ble.py:79: \
FutureWarning: This method will be removed in a future version of \
Bleak. Use the detection_callback of the BleakScanner constructor \
instead.
  scanner.register_detection_callback(detection_callback)

This warning is fixed by moving the callback function to the
BleakScanner constructor.